Make the transition to gcc-elf
Tie off most of the loose ends left by ripping out the binary
modules. Numerous small changes made, but many more are needed to
get rid of all warnings. The firmware seem to work okayish, though.

Bluetooth should be relatively easy to reimplement, using the free
version of BTstack, instead of Kreyos' probably commercial license.

Motion sensing would probably need more work, something able to use
MPU-6050 other than InvenSense's proprietary code would have to be
done.

No bluetooth, no motion sensing in here. Consider yourself warned.

Issue #1

+++ Watch/watch/analog-watch.c
@@ -392,7 +392,7 @@ uint8_t analogclock_process(uint8_t ev, uint16_t lparam, void* rparam)
     if (rparam == NULL)
       selection = window_readconfig()->analog_clock;
     else
-      selection = (uint8_t)rparam - 0x1;
+      selection = (intptr_t)rparam - 0x1;
     if (selection > sizeof(FaceSelections)/sizeof(struct clock_draw) - 1)
       selection = sizeof(FaceSelections)/sizeof(struct clock_draw) - 1;
